fix, server: ran some minor changes - relocated dashboard tool service - updated database script - fixed project repository to actually retrieve top5 projects by end date instead start date
